feat: add `resultWithContext` extension function
This commit introduces a new suspend extension function `resultWithContext` in the `core` module.

This function simplifies executing a suspend block within a specified `CoroutineContext` and wrapping the result (or any thrown exception) in a `Result` type.

It leverages `runCatching` to handle potential exceptions during the execution of the block with `withContext`.